Modify Notes Settings

+++ app/src/main/java/inmethod/gitnotetaking/db/MyDbHelper.kt
@@ -18,12 +18,11 @@ class MyDbHelper(context: Context, name: String,
     override fun onUpgrade(db: SQLiteDatabase,
                            oldVersion:
                            Int, newVersion: Int) {
-        // 刪除原有的表格
-        // 待會再回來完成它
 
-        // 呼叫onCreate建立新版的表格
-        db.execSQL("DROP TABLE IF EXISTS " + RemoteGitDAO.TABLE_NAME)
-        onCreate(db)
+        if(oldVersion == 1 && newVersion == 2) {
+            db.execSQL("DROP TABLE IF EXISTS " + RemoteGitDAO.TABLE_NAME)
+            onCreate(db)
+        }
     }
 
     companion object {
@@ -31,7 +30,7 @@ class MyDbHelper(context: Context, name: String,
         // 資料庫名稱
         val DATABASE_NAME = "gitnotetaking.db"
         // 資料庫版本，資料結構改變的時候要更改這個數字，通常是加一
-        val VERSION = 1
+        val VERSION = 2
 
         // 需要資料庫的元件呼叫這個函式，這個函式在一般的應用都不需要修改
         fun getDatabase(context: Context): SQLiteDatabase {
